test(sdk/recall): move category tests out of history package
TestAllCategories, TestRegisterCategory and TestAllCategoryStrings
exclusively exercise recall.Category APIs but were living in
sdk/history/history_test.go, which forced an otherwise unnecessary
history -> recall test-only dependency. Move them to a new
sdk/recall/categories_test.go (package recall_test, matching the
other black-box tests in the package) and drop the history import.

Also drops a dead code block in the original TestAllCategories that
built an "expected" map and only assigned its values to _ without
asserting on them.

Made-with: Cursor
